routes/auth_routes.py
from flask import Blueprint, request, jsonify
from models import User
from db_config import db
from utils import hash_password, verify_password

bp = Blueprint('auth_routes', __name__)

@bp.route('/register', methods=['POST'])
def register():
    data = request.get_json()
    username = data.get("username")
    password = data.get("password")

    if not username or not password:
        return jsonify({"error": "Missing username or password"}), 400

    if User.query.filter_by(username=username).first():
        return jsonify({"error": "User already exists"}), 400

    new_user = User(username=username, password=hash_password(password))
    db.session.add(new_user)
    db.session.commit()

    return jsonify({"message": "User registered successfully"}), 201

@bp.route('/login', methods=['POST'])
def login():
    data = request.get_json()
    username = data.get("username")
    password = data.get("password")

    if not username or not password:
        return jsonify({"error": "Missing username or password"}), 400

    user = User.query.filter_by(username=username).first()
    if user and verify_password(password, user.password):
        return jsonify({"message": f"Welcome back, {username}!"}), 200

    return jsonify({"error": "Invalid credentials"}), 401
